merge v3.4.x to feature-sync

+++ src/ui/src/views/model-manage/index.vue
@@ -1,27 +1,51 @@
 <template>
     <div class="group-wrapper">
         <cmdb-main-inject
+            :style="{ 'padding-top': showFeatureTips ? '10px' : '' }"
             inject-type="prepend"
             :class="['btn-group', 'clearfix', { sticky: !!scrollTop }]">
+            <feature-tips
+                :feature-name="'model'"
+                :show-tips="showFeatureTips"
+                :desc="$t('ModelManagement[\'模型顶部提示\']')"
+                :more-href="'https://docs.bk.tencent.com/cmdb/Introduction.html#ModelManagement'"
+                @close-tips="showFeatureTips = false">
+            </feature-tips>
             <div class="fl">
-                <bk-button type="primary"
-                    v-if="isAdminView"
-                    :disabled="!$isAuthorized(OPERATION.C_MODEL) || modelType === 'disabled'"
-                    @click="showModelDialog(false)">
-                    {{$t('ModelManagement["新建模型"]')}}
-                </bk-button>
-                <bk-button type="primary"
-                    v-else
-                    v-tooltip="$t('ModelManagement[\'新增模型提示\']')"
-                    :disabled="!$isAuthorized(OPERATION.C_MODEL) || modelType === 'disabled'"
-                    @click="showModelDialog(false)">
-                    {{$t('ModelManagement["新建模型"]')}}
-                </bk-button>
-                <bk-button type="default"
-                    :disabled="!$isAuthorized(OPERATION.C_MODEL_GROUP) || modelType === 'disabled'"
-                    @click="showGroupDialog(false)">
-                    {{$t('ModelManagement["新建分组"]')}}
-                </bk-button>
+                <span v-if="isAdminView" style="display: inline-block;"
+                    v-cursor="{
+                        active: !$isAuthorized($OPERATION.C_MODEL),
+                        auth: [$OPERATION.C_MODEL]
+                    }">
+                    <bk-button type="primary"
+                        :disabled="!$isAuthorized($OPERATION.C_MODEL) || modelType === 'disabled'"
+                        @click="showModelDialog(false)">
+                        {{$t('ModelManagement["新建模型"]')}}
+                    </bk-button>
+                </span>
+                <span v-else style="display: inline-block;"
+                    v-cursor="{
+                        active: !$isAuthorized($OPERATION.C_MODEL),
+                        auth: [$OPERATION.C_MODEL]
+                    }">
+                    <bk-button type="primary"
+                        v-tooltip="$t('ModelManagement[\'新增模型提示\']')"
+                        :disabled="!$isAuthorized($OPERATION.C_MODEL) || modelType === 'disabled'"
+                        @click="showModelDialog(false)">
+                        {{$t('ModelManagement["新建模型"]')}}
+                    </bk-button>
+                </span>
+                <span style="display: inline-block;"
+                    v-cursor="{
+                        active: !$isAuthorized($OPERATION.C_MODEL_GROUP),
+                        auth: [$OPERATION.C_MODEL_GROUP]
+                    }">
+                    <bk-button type="default"
+                        :disabled="!$isAuthorized($OPERATION.C_MODEL_GROUP) || modelType === 'disabled'"
+                        @click="showGroupDialog(false)">
+                        {{$t('ModelManagement["新建分组"]')}}
+                    </bk-button>
+                </span>
             </div>
             <div class="model-type-options fr">
                 <bk-button class="model-type-button enable"
@@ -62,11 +86,19 @@
                     <span class="number">({{classification['bk_objects'].length}})</span>
                     <template v-if="isEditable(classification)">
                         <i class="icon-cc-edit text-primary"
-                            v-if="$isAuthorized(OPERATION.U_MODEL_GROUP)"
+                            :style="{ color: $isAuthorized($OPERATION.U_MODEL_GROUP) ? '' : '#e6e6e6 !important' }"
+                            v-cursor="{
+                                active: !$isAuthorized($OPERATION.U_MODEL_GROUP),
+                                auth: [$OPERATION.U_MODEL_GROUP]
+                            }"
                             @click="showGroupDialog(true, classification)">
                         </i>
                         <i class="icon-cc-del text-primary"
-                            v-if="$isAuthorized(OPERATION.D_MODEL_GROUP)"
+                            :style="{ color: $isAuthorized($OPERATION.D_MODEL_GROUP) ? '' : '#e6e6e6 !important' }"
+                            v-cursor="{
+                                active: !$isAuthorized($OPERATION.D_MODEL_GROUP),
+                                auth: [$OPERATION.D_MODEL_GROUP]
+                            }"
                             @click="deleteGroup(classification)">
                         </i>
                     </template>
@@ -150,19 +182,20 @@
 <script>
     import cmdbMainInject from '@/components/layout/main-inject'
     import theCreateModel from '@/components/model-manage/_create-model'
+    import featureTips from '@/components/feature-tips/index'
     // import theModel from './children'
     import { mapGetters, mapMutations, mapActions } from 'vuex'
     import { addMainScrollListener, removeMainScrollListener } from '@/utils/main-scroller'
-    import { OPERATION } from './router.config.js'
     export default {
         components: {
             // theModel,
             theCreateModel,
-            cmdbMainInject
+            cmdbMainInject,
+            featureTips
         },
         data () {
             return {
-                OPERATION,
+                showFeatureTips: false,
                 scrollHandler: null,
                 scrollTop: 0,
                 groupDialog: {
@@ -182,7 +215,7 @@
             }
         },
         computed: {
-            ...mapGetters(['supplierAccount', 'userName', 'admin', 'isAdminView', 'isBusinessSelected']),
+            ...mapGetters(['supplierAccount', 'userName', 'admin', 'isAdminView', 'isBusinessSelected', 'featureTipsParams']),
             ...mapGetters('objectModelClassify', [
                 'classifications'
             ]),
@@ -226,6 +259,7 @@
             this.searchClassificationsObjects({
                 params: this.$injectMetadata()
             })
+            this.showFeatureTips = this.featureTipsParams['model']
         },
         beforeDestroy () {
             removeMainScrollListener(this.scrollHandler)
@@ -258,12 +292,14 @@
             },
             showGroupDialog (isEdit, group) {
                 if (isEdit) {
+                    if (!this.$isAuthorized(this.$OPERATION.U_MODEL_GROUP)) return
                     this.groupDialog.data.id = group.id
                     this.groupDialog.title = this.$t('ModelManagement["编辑分组"]')
                     this.groupDialog.data.bk_classification_id = group['bk_classification_id']
                     this.groupDialog.data.bk_classification_name = group['bk_classification_name']
                     this.groupDialog.data.id = group.id
                 } else {
+                    if (!this.$isAuthorized(this.$OPERATION.C_MODEL_GROUP)) return
                     this.groupDialog.title = this.$t('ModelManagement["新建分组"]')
                     this.groupDialog.data.bk_classification_id = ''
                     this.groupDialog.data.bk_classification_name = ''
@@ -309,11 +345,17 @@
                 this.hideGroupDialog()
             },
             deleteGroup (group) {
+                if (!this.$isAuthorized(this.$OPERATION.D_MODEL_GROUP)) return
                 this.$bkInfo({
                     title: this.$t('ModelManagement["确认要删除此分组"]'),
                     confirmFn: async () => {
                         await this.deleteClassification({
-                            id: group.id
+                            id: group.id,
+                            config: {
+                                data: this.$injectMetadata({}, {
+                                    inject: !!this.$tools.getMetadataBiz(group)
+                                })
+                            }
                         })
                         this.$store.commit('objectModelClassify/deleteClassify', group['bk_classification_id'])
                     }
@@ -356,7 +398,7 @@
 
 <style lang="scss" scoped>
     .group-wrapper {
-        padding: 76px 20px 20px 0;
+        padding: 130px 20px 20px 0;
     }
     .btn-group {
         position: absolute;
